requirements.txt
pandas
Faker
mysql-connector-python
python-dotenv
flask
flask-cors
matplotlib
numpy
fastapi
uvicorn
langchain
langchain-openai
langchain_community
langchain_experimental
transformers
torch
huggingface_hub
pydantic
crewai
llama-index
sqlalchemy
langchain-groq
llama-index-llms-groq
ratelimit
joblib
tenacity
streamlit
# apache_superset